refact: remove duplicate code in download/install #235
By encapsulate common download functions to achive this.
