+++ src/cli/node/install/install.js
@@ -1,213 +1,16 @@
 /* eslint-disable no-restricted-syntax */
 /* eslint-disable consistent-return */
-const axios = require('axios');
 const chalk = require('chalk');
 const fs = require('fs');
-const fsExtra = require('fs-extra');
-const path = require('path');
-const os = require('os');
-const shell = require('shelljs');
-const semver = require('semver');
-const tar = require('tar');
-const URL = require('url');
 const inquirer = require('inquirer');
 const { spawn } = require('child_process');
-const { symbols, hr, getSpinner, getProgress } = require('core/ui');
-const {
-  print,
-  printError,
-  printInfo,
-  printSuccess,
-  printWarning,
-  fetchAssetsByVersion,
-} = require('core/util');
+const { print, printError, printInfo, printSuccess } = require('core/util');
 const { getPlatform } = require('core/env');
-const debug = require('core/debug')('install');
-const { printLogo } = require('core/util');
-const {
-  getGlobalForgeVersion,
-  getAllChainNames,
-  updateReleaseYaml,
-  isReleaseBinExists,
-  getReleaseAssets,
-  getReleaseDirectory,
-} = require('core/forge-fs');
+const { getAllChainNames, updateReleaseYaml } = require('core/forge-fs');
 const { isForgeStarted } = require('core/forge-process');
+const { formatVersion, downloadForge, DOWNLOAD_FLAGS } = require('../../release/download/lib');
 
-const { DEFAULT_MIRROR, RELEASE_ASSETS, REQUIRED_DIRS } = require('../../../constant');
-
-function fetchReleaseVersion({ mirror, releaseDir }) {
-  if (releaseDir && fs.existsSync(releaseDir)) {
-    const versions = fs
-      .readdirSync(releaseDir)
-      .filter(x => semver.valid(x))
-      .sort((a, b) => {
-        if (semver.gt(a, b)) {
-          return -1;
-        }
-        if (semver.lt(b, a)) {
-          return 1;
-        }
-
-        return 0;
-      });
-
-    if (versions.length) {
-      printSuccess(`Latest forge release version: v${versions[0]}`);
-      return versions[0];
-    }
-  }
-
-  const spinner = getSpinner('Fetching forge release version...');
-  spinner.start();
-
-  try {
-    const url = `${mirror}/forge/latest.json`;
-    const { code, stdout, stderr } = shell.exec(`curl "${url}"`, { silent: true });
-    // debug('fetchReleaseVersion', { code, stdout, stderr, url });
-    if (code === 0) {
-      const { latest: version } = JSON.parse(stdout.trim()) || {};
-      spinner.succeed(`Latest forge release version: v${version}`);
-      return version;
-    }
-    spinner.fail(`Release version fetch error: ${stderr}`);
-  } catch (err) {
-    spinner.fail(`Release version fetch error: ${err.message}`);
-  }
-
-  process.exit(1);
-}
-
-function clearLocalAssets(assets = [], version) {
-  if (assets.length > 0) {
-    printWarning('Cleaning local assets...');
-    assets.forEach(asset => {
-      const assetPath = getReleaseDirectory(asset, version);
-      fsExtra.removeSync(assetPath);
-      debug('cleaned local asset', assetPath);
-    });
-    printSuccess('Cleaned local assets!');
-  }
-}
-
-function getAssetInfo({ platform, version, key, mirror, releaseDir }) {
-  const name = `${key}_${platform}_amd64.tgz`;
-
-  if (releaseDir) {
-    const assetPath = path.join(releaseDir, version, name);
-    if (fs.existsSync(assetPath)) {
-      printSuccess(`Release asset find ${assetPath}`);
-      return { name, url: assetPath };
-    }
-  }
-
-  const url = URL.resolve(mirror, `forge/${version}/${name}`);
-
-  return { url, name };
-}
-
-/**
- *
- * @param {array} assets to be downloaded assets
- * @param {object} download meta info
- */
-async function downloadAssets(assets, { platform, version, mirror, releaseDir }) {
-  const downloadFailedQueue = [];
-
-  // Start download and unzip
-  for (const asset of assets) {
-    try {
-      const assetInfo = getAssetInfo({ platform, version, key: asset, mirror, releaseDir });
-      // eslint-disable-next-line no-await-in-loop
-      const assetTarball = await download(assetInfo);
-      // eslint-disable-next-line no-await-in-loop
-      await expandReleaseTarball(assetTarball, asset, version);
-      fsExtra.removeSync(assetTarball);
-    } catch (error) {
-      printError(error);
-      downloadFailedQueue.push(asset);
-    }
-  }
-
-  if (downloadFailedQueue.length > 0) {
-    print();
-    printError('Failed to download:');
-    downloadFailedQueue.forEach(x => printInfo(x));
-    return false;
-  }
-
-  return true;
-}
-
-function download(assetInfo) {
-  return new Promise((resolve, reject) => {
-    debug('Download asset', assetInfo.url);
-
-    const assetDest = path.join(os.tmpdir(), assetInfo.name);
-    if (fs.existsSync(assetDest)) {
-      shell.rm(assetDest);
-    }
-
-    if (fs.existsSync(assetInfo.url)) {
-      fsExtra.copySync(assetInfo.url, assetDest);
-      printSuccess(`Copied release asset ${assetInfo.url}`);
-      return resolve(assetDest);
-    }
-
-    printInfo(`Start download ${assetInfo.url}`);
-
-    axios
-      .get(assetInfo.url, {
-        responseType: 'stream',
-        timeout: 5 * 60 * 1000, // 5 minutes
-      })
-      .then(response => {
-        const progress = getProgress({
-          title: `${symbols.info} Downloading ${assetInfo.name}`,
-          unit: 'MB',
-        });
-
-        const totalSize = Number(response.headers['content-length']);
-        const bufferArray = [];
-        progress.start((totalSize / 1024 / 1024).toFixed(2), 0);
-
-        let downloadedLength = 0;
-        response.data.on('data', data => {
-          bufferArray.push(data);
-          downloadedLength += Buffer.byteLength(data);
-          progress.update((downloadedLength / 1024 / 1024).toFixed(2));
-        });
-
-        response.data.on('end', () => {
-          if (!response.data.complete) {
-            return reject(new Error('download incomplete'));
-          }
-
-          const buffer = Buffer.concat(bufferArray);
-          fs.writeFileSync(assetDest, buffer);
-          debug(`${assetInfo.name} download success: ${assetDest}`);
-          progress.stop();
-          return resolve(assetDest);
-        });
-
-        response.data.on('error', err => {
-          progress.stop();
-          reject(err);
-        });
-      })
-      .catch(reject);
-  });
-}
-
-async function expandReleaseTarball(filePath, subFolder, version) {
-  const targetDir = path.join(REQUIRED_DIRS.release, subFolder, version);
-
-  fsExtra.removeSync(targetDir); // ensure target directory is empty
-  fs.mkdirSync(targetDir, { recursive: true });
-
-  tar.x({ file: filePath, C: targetDir, strip: 1, sync: true });
-  debug(`Expand release asset ${filePath} to ${targetDir}`);
-}
+const { DEFAULT_MIRROR, RELEASE_ASSETS } = require('../../../constant');
 
 async function main({
   args: [userVersion],
@@ -224,33 +27,6 @@ async function main({
       printInfo(`${chalk.yellow(`Using local releaseDir: ${releaseDir}`)}`);
     }
 
-    const userVer = semver.coerce(userVersion) ? semver.coerce(userVersion).version : '';
-    const version = userVer || fetchReleaseVersion({ mirror, releaseDir });
-
-    if (force === true) {
-      clearLocalAssets(getReleaseAssets(), version);
-    }
-
-    const currentVersion = getGlobalForgeVersion();
-    const versionAssets = await fetchAssetsByVersion(version, platform);
-    // prettier-ignore
-    const unDownloadAssets = RELEASE_ASSETS
-      .filter(x => versionAssets.some(a => a.indexOf(x) >= 0))
-      .filter(x => !isReleaseBinExists(x, version));
-
-    if (unDownloadAssets.length === 0) {
-      printInfo(`forge v${version} is already installed`);
-
-      if (semver.eq(version, currentVersion)) {
-        shell.echo(hr);
-        shell.echo(chalk.cyan('Current forge release'));
-        shell.echo(hr);
-        shell.exec('forge version');
-      }
-
-      return process.exit(0);
-    }
-
     // Ensure forge is stopped, because init on an running node may cause some mess
     if (allowMultiChain === false) {
       const isStarted = await isForgeStarted();
@@ -261,22 +37,23 @@ async function main({
       }
     }
 
-    printLogo();
-
-    const isSuccess = await downloadAssets(unDownloadAssets, {
-      platform,
+    const { version, isLatest } = formatVersion({ version: userVersion, mirror, releaseDir });
+    const downloadResult = await downloadForge({
       version,
       mirror,
       releaseDir,
+      platform,
+      whitelistAssets: RELEASE_ASSETS,
+      force,
+      isLatest,
     });
+
+    if (downloadResult !== DOWNLOAD_FLAGS.SUCCESS) {
+      process.exit(1);
+    }
 
     updateReleaseYaml('forge', version);
     updateReleaseYaml('simulator', version);
-
-    if (!isSuccess) {
-      printError('Please check your assets version or mirror address is correct and try again.');
-      process.exit(1);
-    }
 
     printSuccess(`Congratulations! forge v${version} installed successfully!`);
     print();
@@ -306,19 +83,17 @@ async function main({
         });
 
         childProcess.on('close', () => {
-          shell.echo('');
-          shell.echo(
-            `Configured! Now you can start a forge node with ${chalk.cyan('forge start')}`
-          );
-          shell.echo('');
+          print();
+          print(`Configured! Now you can start a forge node with ${chalk.cyan('forge start')}`);
+          print();
           process.exit(0);
         });
 
         return;
       }
     }
-    shell.echo(`Now you can start a forge node with ${chalk.cyan('forge start')}`);
-    shell.echo('');
+    print(`Now you can start a forge node with ${chalk.cyan('forge start')}`);
+    print();
   } catch (err) {
     printError(err);
     printError('Forge initialize failed, please try again later');
@@ -327,9 +102,3 @@ async function main({
 
 exports.run = main;
 exports.execute = main;
-exports.fetchReleaseVersion = fetchReleaseVersion;
-exports.getAssetInfo = getAssetInfo;
-exports.downloadAssets = downloadAssets;
-exports.download = download;
-exports.expandReleaseTarball = expandReleaseTarball;
-exports.clearLocalAssets = clearLocalAssets;
